requirements.txt
cloudpickle==1.6.0
ConfigSpace==0.3.10
flake8==3.8.4
gym==0.17.3
hpbandster==0.7.4
matplotlib==3.3.4
numba==0.51.2
numpy==1.19.3
numpy-quaternion @ file:///D:/bld/quaternion_1602531791931/work
omegaconf==2.0.3
opencv-python==4.4.0.44
pandas==1.1.3
Pillow @ file:///C:/ci/pillow_1603823068645/work
pybullet==3.0.6
pytorch-lightning==1.1.7
segmentation-models-pytorch==0.1.3
tacto==0.0.2
tensorboard==2.3.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.56.0
hydra-core
hydra_colorlog
sklearn
pypng